+++ template/src/articles/sample/sample.re
@@ -143,25 +143,14 @@ hogehoge
 //}
 これは正式な文章の代わりに入れて使うダミーテキストです。
 
-== Pandoc
+== Markdown
 これは正式な文章の代わりに入れて使うダミーテキストです。
-//pandoc[markdown]{
+//markdown{
 1. これは
     * Markdown
     * です
 2. 出力は
     * LaTeXまたはHTMLで適切に行われます
-//}
-これは正式な文章の代わりに入れて使うダミーテキストです。
-//pandoc[rst]{
-+------------------------+------------+----------+-------------+
-| Header row, column 1   | Header 2   | Header 3 | Header 4    |
-| (header rows optional) |            |          |             |
-+========================+============+==========+=============+
-| body row 1, column 1   | column 2   | column 3 | column 4    |
-+------------------------+------------+----------+-------------+
-| body row 2             | ...        | reStructuredText Table |
-+------------------------+------------+----------+-------------+
 //}
 これは正式な文章の代わりに入れて使うダミーテキストです。
