+++ template/src/layouts/layout.tex.erb
@@ -40,7 +40,6 @@
 \usepackage{afterpage}
 \usepackage{ifoddpage}
 \usepackage{mathrsfs}
-\usepackage{longtable}
 \usepackage[labelformat=simple,labelfont=sf,textfont=sf]{subfig}
 \usepackage{marginfix}
 \usepackage{layout}
